requirements.txt
# 核心依赖
pytest>=7.0.0
requests>=2.28.0
beautifulsoup4>=4.9.3
pyyaml>=6.0.1 